+++ CodeWars/Mathematics/PerimeterOfSquaresInARectangle/PerimeterOfSquaresInARectangle.rs
@@ -1,0 +1,39 @@
+
+struct Fibonacci {
+    curr: u64,
+    next: u64,
+}
+
+impl Fibonacci {
+    pub fn new() -> Self {
+        Fibonacci { curr: 0, next: 1 }
+    }
+}
+
+impl Iterator for Fibonacci {
+    type Item = u64;
+
+    fn next(&mut self) -> Option<u64> {
+        let new_next = self.curr + self.next;
+
+        self.curr = self.next;
+        self.next = new_next;
+
+        Some(self.curr)
+    }
+}
+
+fn perimeter(n: u64) -> u64 {
+    Fibonacci::new().take((n+1) as usize).sum::<u64>() * 4
+}
+
+fn dotest(n: u64, exp: u64) -> () {
+    assert_eq!(perimeter(n), exp)
+}
+
+fn main() {
+    dotest(5, 80);
+    dotest(7, 216);
+    dotest(20, 114624);
+    dotest(30, 14098308);
+}